Refactored next/previous component to follow style standards
